# Remove extraneous allocations.