## What's in this PR?

I removed extraneous allocations to illustrate the techniques for use in other areas of the library if the author would like.

## TODOs

I saw the CLA after I made the changes, my mistake. I'll look into it tomorrow evening maybe, depending how much information I need to disclose.

> benchmark                          old ns/op     new ns/op     delta
> BenchmarkDecrypt/Sample_#1-24      45900         30815         -32.86%
> BenchmarkDecrypt/Sample_#2-24      45662         31572         -30.86%
> BenchmarkDecrypt/Sample_#3-24      49453         39011         -21.11%
> BenchmarkDecrypt/Sample_#4-24      44523         37762         -15.19%
> BenchmarkDecrypt/Sample_#5-24      40971         25624         -37.46%
> BenchmarkDecrypt/Sample_#6-24      45525         27824         -38.88%
> BenchmarkDecrypt/Sample_#7-24      42651         30249         -29.08%
> BenchmarkDecrypt/Sample_#8-24      49264         39506         -19.81%
> BenchmarkDecrypt/Sample_#9-24      42568         31181         -26.75%
> BenchmarkDecrypt/Sample_#10-24     42259         25162         -40.46%
> BenchmarkDecrypt/Sample_#11-24     46210         32699         -29.24%
> BenchmarkDecrypt/Sample_#12-24     39739         29340         -26.17%
> BenchmarkDecrypt/Sample_#13-24     50127         34662         -30.85%
> BenchmarkDecrypt/Sample_#14-24     52410         37990         -27.51%
> BenchmarkDecrypt/Sample_#15-24     44501         30777         -30.84%
> 
> benchmark                          old allocs     new allocs     delta
> BenchmarkDecrypt/Sample_#1-24      231            102            -55.84%
> BenchmarkDecrypt/Sample_#2-24      237            108            -54.43%
> BenchmarkDecrypt/Sample_#3-24      234            105            -55.13%
> BenchmarkDecrypt/Sample_#4-24      234            105            -55.13%
> BenchmarkDecrypt/Sample_#5-24      228            99             -56.58%
> BenchmarkDecrypt/Sample_#6-24      237            108            -54.43%
> BenchmarkDecrypt/Sample_#7-24      231            102            -55.84%
> BenchmarkDecrypt/Sample_#8-24      234            105            -55.13%
> BenchmarkDecrypt/Sample_#9-24      234            105            -55.13%
> BenchmarkDecrypt/Sample_#10-24     228            99             -56.58%
> BenchmarkDecrypt/Sample_#11-24     231            102            -55.84%
> BenchmarkDecrypt/Sample_#12-24     234            105            -55.13%
> BenchmarkDecrypt/Sample_#13-24     237            108            -54.43%
> BenchmarkDecrypt/Sample_#14-24     237            108            -54.43%
> BenchmarkDecrypt/Sample_#15-24     228            99             -56.58%
> 
> benchmark                          old bytes     new bytes     delta
> BenchmarkDecrypt/Sample_#1-24      7728          2912          -62.32%
> BenchmarkDecrypt/Sample_#2-24      7728          2928          -62.11%
> BenchmarkDecrypt/Sample_#3-24      8176          3344          -59.10%
> BenchmarkDecrypt/Sample_#4-24      8176          3344          -59.10%
> BenchmarkDecrypt/Sample_#5-24      7696          2912          -62.16%
> BenchmarkDecrypt/Sample_#6-24      7760          2928          -62.27%
> BenchmarkDecrypt/Sample_#7-24      7728          2912          -62.32%
> BenchmarkDecrypt/Sample_#8-24      8176          3344          -59.10%
> BenchmarkDecrypt/Sample_#9-24      8176          3344          -59.10%
> BenchmarkDecrypt/Sample_#10-24     7696          2912          -62.16%
> BenchmarkDecrypt/Sample_#11-24     7728          2912          -62.32%
> BenchmarkDecrypt/Sample_#12-24     7744          2928          -62.19%
> BenchmarkDecrypt/Sample_#13-24     8192          3360          -58.98%
> BenchmarkDecrypt/Sample_#14-24     8192          3360          -58.98%
> BenchmarkDecrypt/Sample_#15-24     7696          2912          -62.16%
>