@anitgandhi 2017-07-06
Thanks for the informative PR! 

I do request that you sign the CLA if you'd like this merged. If not, would it be ok if I piggy-backed on this code and applied it to the other package as well?