@cstockton 2017-07-07
I signed the individual CLA, though I didn't (and won't) provide a cell phone number or address. Let me know if you have any questions though I would be glad to help.