@cstockton 2017-07-25
@jaredsmith What is your privacy policy around this personal information? It's a google doc which makes it very unclear who has access to these details. While sometimes I come across a CLA that may ask for my address, It's unusual to require a personal phone number in a CLA, under what scenario would CapitalOne even need my address- let alone call me?