@jaredsmith 2017-08-02
That's a great question -- we'll only contact you related to your open source contributions -- we don't use the data for any other reasons.  Our legal team just likes having an address and phone number if there's a legal question about a particular contribution.  The only people who have access to the data are the project leaders (so that they can see if someone has signed a CLA or not) and our Open Source Office.  Nobody else has the data.

If there's any other questions I can answer regarding the CLAs or our open source office, please don't hesitate to ask.