@cstockton 2017-08-02
@jaredsmith I'll sign it after work as long as it will not be shared with third parties, used for marketing, advertising or recruitment purposes. That is: Only used to contact me in the event a legal dispute or question on code I've contributed. Though I would appreciate being tagged here first. Maybe a little clarity on this on the document would be helpful. Thanks!